add shop frontend and update privacy visa frontend
