+++ privacy_visa/front/package.json
@@ -13,6 +13,7 @@
     "circomlib": "^2.0.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.27.0",
     "react-scripts": "5.0.1",
     "snarkjs": "^0.7.4",
     "wasm-loader": "^1.3.0",
